+++ app/commands/timeCommands.py
@@ -55,7 +55,36 @@ def load_csv(csv_file):
         print("Dados inseridos")
     else:
         exit()
+
+@timeCommands.cli.command("loadEmojis")
+@click.argument("csv_file")
+def load_emoji(csv_file):
+    data = []
+    with open(csv_file) as arq:
+        headers = next(arq, None)
+        # Caso valor default None, retorna []
+        if headers is None:
+            return []
+
+        cabecalho = headers.strip().split(SEPARADOR_CSV)
+        #print(cabecalho)
+
+        for linha in arq:
+            #print("linha:",linha)
+            colunas = linha.strip().split(SEPARADOR_CSV)
+            documento = zip(cabecalho,colunas)
+            documento = dict(documento)
+            data.append(documento)
+
+        if not data:
+            print("Sem dados.")
+            return None
+        
+        print(data)
     
+    timeCollection = mongo.db.emoji
+    timeCollection.insert(data)
+
 
 @timeCommands.cli.command("getRank")
 @click.argument("rank")