+++ CHANGELOG.md
@@ -4,8 +4,15 @@ All notable changes to this project will be documented in this file. Based on [K
 
 ## [Unreleased]
 
+
+## [1.0.1] - 2021-09-09
+
 ### Fixed
 - Tables page layout
+
+### Added
+- Emoji flags database
+- Command line function to report last games and users who did not bet yet
 
 ## [1.0.0] - 2021-08-21
 
@@ -34,7 +41,8 @@ All notable changes to this project will be documented in this file. Based on [K
 - Minimal ranking page
 - Table from games from first season year 20
 
-[unreleased]: https://github.com/leafarlins/gokopa/compare/v1.0.0...HEAD
+[unreleased]: https://github.com/leafarlins/gokopa/compare/v1.0.1...HEAD
+[1.0.1]: https://github.com/leafarlins/gokopa/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/leafarlins/gokopa/compare/v0.2.0...v1.0.0
 [0.2.0]: https://github.com/leafarlins/gokopa/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/leafarlins/gokopa/releases/tag/v0.1.0